Transfer executable scripts into actions and script configuration parameters into config or command line
