+++ darepy/config/cfg.py
@@ -7,9 +7,13 @@ For practicle purposes this behavior makes the module behave similar to a Single
 
 from .experiment import Experiment
 from .instruments import SANS1Cfg, SANSLLBCfg
+from .reduction import Reduction, Merging
 
 experiment = Experiment.load()
+reduction = Reduction.load()
+merging = Merging.load()
 
 sans1 = SANS1Cfg.load()
 sansllb = SANSLLBCfg.load()
 
+instrument=sansllb
