+++ darepy/post_processing.py
@@ -9,11 +9,11 @@ Created on Wed Jul 13 20:50:57 2022
 import numpy as np
 import matplotlib.pyplot as plt
 import pickle
-import integration as integ
+from . import integration as integ
 from scipy.interpolate import interp1d
 import scipy.optimize
 import os
-from utils import smooth
+from .utils import smooth
 from scipy.stats import linregress # To check for linearity/constancy
 
 # %% plot_all_data
